Refactor: USER_NOT_FOUND errorcode추가
